Moving `ValueEvaluator` to `cpg-core`
This PR moves `ValueEvaluator` to `cpg-core` and makes it possible for a language to override the default.
