+++ cpg-analysis/src/main/kotlin/de/fraunhofer/aisec/cpg/passes/UnreachableEOGPass.kt
@@ -26,7 +26,6 @@
 package de.fraunhofer.aisec.cpg.passes
 
 import de.fraunhofer.aisec.cpg.TranslationContext
-import de.fraunhofer.aisec.cpg.analysis.ValueEvaluator
 import de.fraunhofer.aisec.cpg.graph.Node
 import de.fraunhofer.aisec.cpg.graph.declarations.FunctionDeclaration
 import de.fraunhofer.aisec.cpg.graph.edges.flows.EvaluationOrder
@@ -47,8 +46,6 @@ import de.fraunhofer.aisec.cpg.passes.configuration.DependsOn
  */
 @DependsOn(ControlFlowSensitiveDFGPass::class)
 open class UnreachableEOGPass(ctx: TranslationContext) : EOGStarterPass(ctx) {
-
-    protected open val evaluator = ValueEvaluator()
 
     override fun cleanup() {
         // Nothing to do
@@ -143,7 +140,7 @@ open class UnreachableEOGPass(ctx: TranslationContext) : EOGStarterPass(ctx) {
         n: IfStatement,
         state: UnreachabilityStateElement,
     ): UnreachabilityStateElement {
-        val evalResult = evaluator.evaluate(n.condition)
+        val evalResult = n.language.evaluator.evaluate(n.condition)
 
         val (unreachableEdges, remainingEdges) =
             if (evalResult == true) {
@@ -192,7 +189,7 @@ open class UnreachableEOGPass(ctx: TranslationContext) : EOGStarterPass(ctx) {
                 is ForStatement -> n.condition
                 else -> return state
             }
-        val evalResult = evaluator.evaluate(condition)
+        val evalResult = n.language.evaluator.evaluate(condition)
 
         val (unreachableEdges, remainingEdges) =
             if (evalResult is Boolean && evalResult == true) {
