+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/evaluation/NumberSet.kt
@@ -23,7 +23,7 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.analysis
+package de.fraunhofer.aisec.cpg.evaluation
 
 import de.fraunhofer.aisec.cpg.graph.Node
 import org.apache.commons.lang3.builder.ToStringBuilder
