+++ cpg-core/src/test/kotlin/de/fraunhofer/aisec/cpg/evaluation/ValueEvaluationTests.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Fraunhofer AISEC. All rights reserved.
+ * Copyright (c) 2025, Fraunhofer AISEC. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,23 +23,23 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.testcases
+package de.fraunhofer.aisec.cpg.evaluation
 
 import de.fraunhofer.aisec.cpg.TranslationConfiguration
 import de.fraunhofer.aisec.cpg.frontends.TestLanguage
 import de.fraunhofer.aisec.cpg.frontends.testFrontend
-import de.fraunhofer.aisec.cpg.graph.*
+import de.fraunhofer.aisec.cpg.graph.array
 import de.fraunhofer.aisec.cpg.graph.builder.*
-import de.fraunhofer.aisec.cpg.passes.UnreachableEOGPass
+import de.fraunhofer.aisec.cpg.graph.newConditionalExpression
+import de.fraunhofer.aisec.cpg.graph.newNewArrayExpression
 
 class ValueEvaluationTests {
     companion object {
         fun getSizeExample(
             config: TranslationConfiguration =
-                TranslationConfiguration.builder()
+                TranslationConfiguration.Companion.builder()
                     .defaultPasses()
                     .registerLanguage<TestLanguage>()
-                    .registerPass<UnreachableEOGPass>()
                     .build()
         ) =
             testFrontend(config).build {
@@ -98,10 +98,9 @@ class ValueEvaluationTests {
 
         fun getComplexExample(
             config: TranslationConfiguration =
-                TranslationConfiguration.builder()
+                TranslationConfiguration.Companion.builder()
                     .defaultPasses()
                     .registerLanguage<TestLanguage>()
-                    .registerPass<UnreachableEOGPass>()
                     .build()
         ) =
             testFrontend(config).build {
@@ -140,10 +139,9 @@ class ValueEvaluationTests {
 
         fun getExample(
             config: TranslationConfiguration =
-                TranslationConfiguration.builder()
+                TranslationConfiguration.Companion.builder()
                     .defaultPasses()
                     .registerLanguage<TestLanguage>()
-                    .registerPass<UnreachableEOGPass>()
                     .build()
         ) =
             testFrontend(config).build {
@@ -245,10 +243,9 @@ class ValueEvaluationTests {
 
         fun getCfExample(
             config: TranslationConfiguration =
-                TranslationConfiguration.builder()
+                TranslationConfiguration.Companion.builder()
                     .defaultPasses()
                     .registerLanguage<TestLanguage>()
-                    .registerPass<UnreachableEOGPass>()
                     .build()
         ) =
             testFrontend(config).build {
